feat(handlers): Reorganised `peers.go` to smaller files since each section will eventually be quite long
